+++ VoiceCraft.Core/Network/Packets/SetMaxRangePacket.cs
@@ -5,10 +5,10 @@ namespace VoiceCraft.Core.Network.Packets
     public class SetMaxRangePacket : VoiceCraftPacket
     {
         public override PacketType PacketType => PacketType.SetMaxRange;
-        public byte Id { get; private set; }
+        public int Id { get; private set; }
         public int MaxRange { get; private set; }
 
-        public SetMaxRangePacket(byte id = 0, int maxRange = 0)
+        public SetMaxRangePacket(int id = 0, int maxRange = 0)
         {
             Id = id;
             MaxRange = maxRange;
@@ -22,7 +22,7 @@ namespace VoiceCraft.Core.Network.Packets
 
         public override void Deserialize(NetDataReader reader)
         {
-            Id = reader.GetByte();
+            Id = reader.GetInt();
             MaxRange = reader.GetInt();
         }
     }
